+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     - Specifications now extend base class
     - Rename Dataset validate argument to verify
     
+- 0.1.3
+    - Optimized Cosine distance kernel
+    - Optimized (NaN) Safe Euclidean distance kernel
+
 - 0.1.2
     - Fixed Grid Search best hyper-parameters method
     - Fixed K Means average loss calculation